Try half-dynamic approach for meta blocks
